+++ sources/packages/backend/apps/api/src/route-controllers/application/application.students.controller.ts
@@ -70,7 +70,7 @@ import { CustomNamedError } from "@sims/utilities";
 import { PrimaryIdentifierAPIOutDTO } from "../models/primary.identifier.dto";
 import { ApplicationData } from "@sims/sims-db/entities/application.model";
 import { ApplicationStatus, StudentAppealStatus } from "@sims/sims-db";
-import { DisbursementScheduleSharedService } from "@sims/services";
+import { ConfirmationOfEnrollmentService } from "@sims/services";
 
 @AllowAuthorizedParty(AuthorizedParties.student)
 @RequiresStudentAccount()
@@ -83,7 +83,7 @@ export class ApplicationStudentsController extends BaseController {
     private readonly studentService: StudentService,
     private readonly programYearService: ProgramYearService,
     private readonly offeringService: EducationProgramOfferingService,
-    private readonly disbursementScheduleSharedService: DisbursementScheduleSharedService,
+    private readonly confirmationOfEnrollmentService: ConfirmationOfEnrollmentService,
     private readonly assessmentService: StudentAssessmentService,
     private readonly applicationControllerService: ApplicationControllerService,
     private readonly craIncomeVerificationService: CRAIncomeVerificationService,
@@ -121,7 +121,7 @@ export class ApplicationStudentsController extends BaseController {
         application.data,
       );
     const firstCOEPromise =
-      this.disbursementScheduleSharedService.getFirstDisbursementScheduleByApplication(
+      this.confirmationOfEnrollmentService.getFirstDisbursementScheduleByApplication(
         applicationId,
       );
     const [applicationData, firstCOE] = await Promise.all([
