+++ sources/packages/backend/apps/api/src/services/disbursement-schedule/_tests_/unit/disbursement-schedule.getCOEApprovalPeriodStatus.spec.ts
@@ -1,29 +1,25 @@
 import {
   ConfirmationOfEnrollmentService,
-  DisbursementScheduleSharedService,
   NotificationActionsService,
   SequenceControlService,
-  SystemUsersService,
 } from "@sims/services";
-import { DataSource } from "typeorm";
+import { DataSource, Repository } from "typeorm";
 import { COE_WINDOW, addDays } from "@sims/utilities";
 import { COEApprovalPeriodStatus } from "../../disbursement-schedule.models";
+import { DisbursementSchedule } from "@sims/sims-db";
 
 describe("DisbursementScheduleService-getCOEApprovalPeriodStatus", () => {
-  let service: DisbursementScheduleSharedService;
+  let service: ConfirmationOfEnrollmentService;
 
   beforeAll(async () => {
     const dataSource = {} as DataSource;
     dataSource.getRepository = jest.fn();
     const sequenceService = {} as SequenceControlService;
+    const disbursementScheduleRepo = {} as Repository<DisbursementSchedule>;
     const notificationActionsService = {} as NotificationActionsService;
-    const confirmationOfEnrollmentService =
-      {} as ConfirmationOfEnrollmentService;
-    const systemUsersService = {} as SystemUsersService;
-    service = new DisbursementScheduleSharedService(
+    service = new ConfirmationOfEnrollmentService(
+      disbursementScheduleRepo,
       dataSource,
-      systemUsersService,
-      confirmationOfEnrollmentService,
       sequenceService,
       notificationActionsService,
     );
