+++ sources/packages/backend/libs/services/src/confirmation-of-enrollment/models/confirmation-of-enrollment.models.ts
@@ -33,3 +33,22 @@ export enum MaxTuitionRemittanceTypes {
    */
   Effective,
 }
+
+/**
+ * COE approval period status.
+ */
+export enum COEApprovalPeriodStatus {
+  /**
+   * COE is within its valid approval period and can be approved.
+   */
+  WithinApprovalPeriod = "Within approval period",
+  /**
+   * COE is currently before its approval period
+   * and needs to wait until approval period to be confirmed.
+   */
+  BeforeApprovalPeriod = "Before approval period",
+  /**
+   * The COE is currently crossed its approval period and cannot be confirmed by institution.
+   */
+  AfterApprovalPeriod = "After approval period",
+}
