+++ sources/packages/backend/libs/services/src/disbursement-schedule/disbursement-schedule.models.ts
@@ -17,22 +17,3 @@ export interface AwardValueWithRelatedSchedule {
   relatedSchedule: ECertDisbursementSchedule;
   awardValue: DisbursementValue;
 }
-
-/**
- * COE approval period status.
- */
-export enum COEApprovalPeriodStatus {
-  /**
-   * COE is within its valid approval period and can be approved.
-   */
-  WithinApprovalPeriod = "Within approval period",
-  /**
-   * COE is currently before its approval period
-   * and needs to wait until approval period to be confirmed.
-   */
-  BeforeApprovalPeriod = "Before approval period",
-  /**
-   * The COE is currently crossed its approval period and cannot be confirmed by institution.
-   */
-  AfterApprovalPeriod = "After approval period",
-}
